refactor auto login: quit depending on LoginViewController login user even if theres no Internet connection
